Move kubernetes types out of common
